refactor: replace MUI Dialog in PasteRecordsButton with window-manager WindowContent
