+++ designer/client/src/windowManager/WindowKind.tsx
@@ -27,4 +27,5 @@ export enum WindowKind {
     aiAssistant,
     enterpriseFeatureInfo,
     saveAsTestCase,
+    pasteRecords,
 }
